Remove cache for resource indexes.
In the case of a library not having an R Class (i.e: null) it breaks the cache since HashMap allows a null key and adding to the map clobbered resource indexes. We could cache by package name but there are cases where projects have multiple dependency libraries with the same package name and this would also clobber resource indexes. The best thing to do is to remove the cache and create all resource indexes since the RoutingResourceLoader merges these anyway.